backend_django/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
channels>=4.1.0
daphne>=4.1.0
channels-redis>=4.2.0
cloudinary>=1.40.0
Pillow>=10.0.0
gunicorn>=22.0.0
whitenoise>=6.7.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
